cli: Make openvdc/cmd/util.go internal package.
