+++ cmd/openvdc/cmd/run.go
@@ -5,6 +5,7 @@ import (
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/axsh/openvdc/api"
+	"github.com/axsh/openvdc/cmd/openvdc/internal/util"
 	"github.com/spf13/cobra"
 	"golang.org/x/net/context"
 	"google.golang.org/grpc"
@@ -25,7 +26,7 @@ var runCmd = &cobra.Command{
 
 		templateSlug := args[0]
 		req := prepareRegisterAPICall(templateSlug)
-		return remoteCall(func(conn *grpc.ClientConn) error {
+		return util.RemoteCall(func(conn *grpc.ClientConn) error {
 			c := api.NewInstanceClient(conn)
 			res, err := c.Run(context.Background(), req)
 			if err != nil {
