+++ cmd/openvdc/cmd/unregister.go
@@ -7,6 +7,7 @@ import (
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/axsh/openvdc/api"
+	"github.com/axsh/openvdc/cmd/openvdc/internal/util"
 	"github.com/spf13/cobra"
 	"google.golang.org/grpc"
 )
@@ -24,7 +25,7 @@ var unregisterCmd = &cobra.Command{
 		if len(resourceID) == 0 {
 			log.Fatalf("Invalid Resource ID: %s", resourceID)
 		}
-		return remoteCall(func(conn *grpc.ClientConn) error {
+		return util.RemoteCall(func(conn *grpc.ClientConn) error {
 			c := api.NewResourceClient(conn)
 			res, err := c.Unregister(context.Background(), &api.ResourceIDRequest{Key: &api.ResourceIDRequest_ID{ID: resourceID}})
 			if err != nil {
